fixes sizing and centering issue with previews
fixes CNVS-18558

When previewing a video or audio file in new files, the video was
showing up really small. The problem was with when height for the
containing video was being calculated. This ensures the container div
retains it's height and width values, the video loads, then css to
center the video is added.

Test Plan
Given you're working in new files
And you have uploaded a video file
When you click on the video file to open a preview
Then you should see the video centered in the preview page
And the video should be large enough to take up most of the space
And clicking the "full screen" button should work

Change-Id: Id2ebbcb19aff2020d4ca4e06775fd2e808b1f9cb
Reviewed-on: https://gerrit.instructure.com/49977
Tested-by: Jenkins
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>