Translate equations back into editor
fixes CNVS-18342

This was treating "img" tags inside the editor
as text for the purposes of extracting an equation.  By
checking the "text" with a regex for actually being an img tag,
in that case we can wrap it in a jquery object and parse it
the right way for the equation editor.

Also put in some specs for the EquationEditorView

TEST PLAN:
- go to a tinymce editor
- Click on the LaTeX "π" icon.
- Create any equation & click "Insert Equation"
- Select the inserted equation and click on the (now highlighted) LaTeX
"π" icon.
- Make sure that the entire HTML img tag doesn't get pulled in, just
  the alt attribute

Change-Id: Ic917af7698c64dc22e07afa7e514e6cb22e6f858
Reviewed-on: https://gerrit.instructure.com/50362
Tested-by: Jenkins
Reviewed-by: Ryan Shaw <ryan@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>