remove the js server error reporting
This data has proven to not be actionable or useful.

We are keeping in the window onerror handler, as the majority of
javascript errors are caused by third party scripts and integrations,
and we want canvas to continue operating when those occur. There was
code to try and detect third party errors based on script url, but that
hasn't worked in practice because the error url often ends up being on
the canvas domain, even when the error is in an external script.

I'll admit this feels a bit like turning a blind eye to potential
issues, and I'd love to revisit JS error reporting in the future, but
from a practical standpoint we just can't sift through these as they are
today, and they make building out a reasonable error reporting
infrastructure much more difficult.

Change-Id: Ib3f035358dd7e43c812724534c2fe80d9a5cc2c2
Reviewed-on: https://gerrit.instructure.com/50877
Tested-by: Jenkins
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Brian Palmer <brianp@instructure.com>