refactor coupling start/end of time ranges
refs CNVS-19515

put the functionality in coupleTimeFields.coffee, and then use it in the
calendar event details. in the process, fix it to use the data attributes on
the fields instead of treating the related datetime_suggest fields as data
(they're not)

for the purpose of being able to set new values on the fields and have them
update their data attributes and suggest values, added setTime and setDate
methods (and setDatetime, though that's not used yet) to the
DatetimeField instance. use this instead of manually triggering change
events after direct accesses to val()

test-plan:
 - regression testing around editing calendar events

Change-Id: I2f403d54b336bcfc4c118a31fb7e3b27129a45f7
Reviewed-on: https://gerrit.instructure.com/53189
Reviewed-by: Ethan Vizitei <evizitei@instructure.com>
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>