OutcomeLink API objects: Add 'assessed' trait
This provides a way to query the learning outcome object to see if it is
assessed in a particular course.

Previously it was only possible to query to see if it had been assessed
anywhere.  This is problematic because we sometimes do care whether the
assessing was done for a particular course (such as deciding whether to
allow the outcome to be removed from a course)

Add course_ids_where_assessed and others to learning outcome model

We have need now to know if a particular outcome has been assessed in
a given course (or context), that is not necessarily the context stored
on the outcome object itself.

For example, when an Account level outcome is imported into a course, it
maintains the account and account ID as its context.  The problem then,
is how do we know whether this outcome has been assessed in the course
in which it has been imported, and in which we are now viewing the
object?

The answer, is to add an 'assessed' trait on the outcome link object
returned by the API.  This outcome link contains information about the
context that we are interested in.  In this example, it is the course in
which it was imported.

Using this new trait, clients (including our UI) can tell if a student
has been assessed with this outcome in the particular context, rather
than just globally.

Refs CNVS-19806

Test Plan:
    - Make an account level outcome
    - import that outcome into two courses
    - assess the outcome in one of the courses
    - from a rails console, call the assessed? method and pass in each
      course object.  For the course that was assessed it should return
      true, and for the other it should return false.
        - Get a learning outcome object
           - LearningOutcome.find(<id)
           - LearningOutcome.where(context_id: <course_id>)
        - Call the assessed? method and pass in the course or course id
            - learning_outcome.assessed?(course.id)
        - To get an array of all courses:
            - Course.all
    - from the API, request the outcome links for the course.
    - ensure that the link for the course in which the outcome was
      assessed has the 'assessed' field set to 'true', and the other has
      it set to 'false'

Change-Id: If9a0260289704944145cc3a0929e1183656dfb68
Reviewed-on: https://gerrit.instructure.com/52509
Tested-by: Jenkins
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Reviewed-by: John Corrigan <jcorrigan@instructure.com>
Product-Review: Benjamin Porter <bporter@instructure.com>