make tabbing through SlickGrid more accessible
This change makes it so you can tab out of the
SlickGrid when you are on the last cell.

This change also makes it so enter takes you to
the top of the next column.

fixes CNVS-18686
refs CNVS-17043

Test Plan:
- Import a CSV into the gradebook, regardless of which one.
- Get to the final results page, where it will ask you to save.
- You should be able to reach the submit/save/etc. buttons
  with only the keyboard.
- Please also check using the test plan for CNVS-17043 since
this commit actually adds this functionality into Canvas:
  - In the gradebook press the enter key while editing the
    bottom row of a column.
  - Focus should go to the top of the next column.

Change-Id: Ie9add59ccc1873708cfd45df7fdc7f8032aa7cb1
Reviewed-on: https://gerrit.instructure.com/52784
Tested-by: Jenkins
Reviewed-by: Jonathan Featherstone <jfeatherstone@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>
Product-Review: Strand McCutchen <smccutchen@instructure.com>