Add 'mark done' module progression constraint
In module based courses, to make it easier for students to keep
track of their progress. This commit adds support for 'mark as
done' requirements to the module progressions. This to better
cater to students who needs multiple reading sessions before they
feel they are done with a module item (i.e wiki pages).

This requirement can also be used on assignments that are purely
for self studying.

Test plan:
  - In a course with modules
  - Make sure there is one wiki page and one assignment associated with a module
  - Add 'mark as done' as module completion requirements to both module items
  - As a student of the course, go to the modules page
  - Oberve that both module items are labeled as "must mark as done"
  - Visit the module items each in turn and click the "mark as done" button.
  - On the modules page, oberve that both module items are now labeled as "marked done".