gradebook column position now saves to db
if an assignment, aggregate or total column is
reordered in the gradebook, the custom ordering
will be saved to the user's preferences. if any
column is resized, that size will be saved to
the user's preferences for that assignment id.
this allows column size and column position
settings to persist between multiple computers
/ browsers.

closes CNVS-17052

Change-Id: Ibe77de372b17764233a7662c3fc8517ddc9df1dc
Reviewed-on: https://gerrit.instructure.com/52017
Tested-by: Jenkins
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Product-Review: Dylan Ross <dross@instructure.com>