spec: fix fragile outcomes specs
For the first one, the outcomes must be being rendered in a way that
`wait_for_ajaximations` is not catching. Wrapping it in
`keep_trying_until` is a blunt but effective way to work around that.

The second problematic spec was relying on a specific ordering of
outcomes to show up in the list, which was not reliable. The change
forces lookups by id so it's consistent.

Change-Id: Ie20ac3c73bab6b48410036f10a70b3432476acd2
Reviewed-on: https://gerrit.instructure.com/53343
Tested-by: Jenkins
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Product-Review: Zach Wily <zach@instructure.com>
QA-Review: Zach Wily <zach@instructure.com>