change text on Email column of confirm-add-users dialog
this column may contain an email or a login, depending on
what the user searched for.  it might not be the same thing
in each row.  make the column header read "Email/Login"
to reflect this.

test plan: add a user to a course and note the column
header in the confirm dialog.

fixes CNVS-8242

Change-Id: I106c3c34b3db5f56555ffdb6204cfaedee1d71b9
Reviewed-on: https://gerrit.instructure.com/53965
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Cosme Salazar <cosme@instructure.com>