verify that grading periods effect editability of assignments
refs CNVS-20459
fixes CNVS-19933

Using these specs I identified an area of the JavaScript
where a null due_at is coerced into a Date object for
comparison.

Given multiple grading periods is enabled,
And there are admin-created current and past grading
  periods,
When a teacher attempts to edit an assignment in the past
  grading period,
Then they should not be able to edit the assignment.
When a teacher attempts to edit an assignment with no due
  date,
Then they should be able to edit the assignment.

The added integration specs verify that:
  - an admin can edit an assignment which is in a past
    grading period.
  - an admin can edit an assignment which has no due_at
  - a teacher cannot edit an assignment which is in a past
    grading period.
  - a teacher can edit an assignment which has no due_at

Change-Id: I4cc249a9ca90572180bea9f0b044e367c6ccef94
Reviewed-on: https://gerrit.instructure.com/55180
Tested-by: Jenkins
Reviewed-by: Dylan Ross <dross@instructure.com>
QA-Review: Robert Lamb <rlamb@instructure.com>
Product-Review: Strand McCutchen <smccutchen@instructure.com>