warn user of unsaved comments in speedgrader
when a user is leaving speedgrader (exiting the window, linking
to another page, or refreshing the page), they will be prompted
with a message if they have unsaved comments on the page. the prompt
will give the user the option to 1) stay on the current page, or
2) continue on to the next page. If all comments have been saved,
however, the page will not prompt the user with a message.

note: firefox versions 4+ will always display the same generic
message: 'This page is asking you to confirm that you want to
leave - data you have entered may not be saved.' this is a
documented feature and there is no way to customize the message
for FF (see 'Notes' section at https://developer.mozilla.org/en-US
/docs/Web/Events/beforeunload)

closes CNVS-10809

test plan:

create a quiz as a teacher, and then log in as a student and take
the quiz. then log back in as a teacher and go to the speedgrader.
enter some text in the 'Add Comment' section on the right-hand side
of the screen but _don't_ click the 'Submit Comment' button. try
navigating away from the page (x out of the window, refresh the
page, or click on a link that takes you to another page) and verify
that a prompt shows up telling you that there are unsaved comments.
this prompt should not show up if you have saved all your comments
before navigating away from the page.

edge cases considered: try entering a comment of ONLY space characters,
and don't click the 'Submit Comment' button. when you navigate away
from the page you should _not_ get a prompt, since a blank comment is
not considered a comment that needs saving.

Change-Id: I198738814cdec9e6ca78a670ba06b15c5a1af836
Reviewed-on: https://gerrit.instructure.com/55553
Tested-by: Jenkins
QA-Review: Adrian Foong <afoong@instructure.com>
Reviewed-by: Strand McCutchen <smccutchen@instructure.com>
Product-Review: Spencer Olson <solson@instructure.com>