Makes focus return to wiki cog after focus
It's slightly hacky, but this makes it so that we pass up a selector
that can be used to set focus to after a render.  This is because
capturing the actual element that focus should go back to is futile
since it gets wiped out before the render.

fixes CNVS-20199
refs CNVS-19340

Test Plan:
  - Go to the wiki pages (/courses/XX/pages)
  - Use the cog menu to edit a page.
  - Rename and save the page
  - The focus should return to the cog menu

Change-Id: Ie71ab324bdfbba54573d0abafa3f18c542ba2f0b
Reviewed-on: https://gerrit.instructure.com/55645
Tested-by: Jenkins
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>