don't manually trigger context_module_action on entry save
fixes CNVS-18588 (maybe)

In this method we save the entry twice (each time triggering a call to
`context_module_action_later`, then also called `context_module_action`
explicitly. I believe this could lead to a situation in which we are
creating and updating the submission associated with the entry twice in
quick succession, thereby updating the submitted_at date and triggering
the resubmit notification.

test plan:
- as a teacher, update your notificatino preferences to receive a
  message for assignment submissions ASAP.
- create a graded discussion topic.
- as a student, reply to the topic.
- as the teacher, look at your messages (/users/:user_id/messages), and
  observe that there is only a notification for the submission, not a
  re-submission.
- Also: regression testing around the creation of discussion entries.

NOTE: This is only an intermittently reproduceable bug that seems to rely
on background jobs running. See jira ticket for an environment in which
the issue was reproduceable with some consistency.

Change-Id: Ic33af0fa5d70f90f2d1ad1b1b1d7ea5cbc7e05dd
Reviewed-on: https://gerrit.instructure.com/56289
Tested-by: Jenkins
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>