Make announcement dates accessible
closes CNVS-19317
refs CNVS-15543

This applies a screenreader-only label to the date picker and
timepickers used in the announcements UI, and some KB-only user tooltips
(though they are visible for all users).  The goal is to make
it clear what format to input dates and times in to users
for whom the rich selector isn't so great.

Also adds some parameters to the accessible helpers to make
them parallel with the ones available in handlebars_helper.js

TEST PLAN:
- go to account settings
- navigate to the tab for announcements
- Using a screenreader with your eyes closed, ensure you can figure out
  what date format to submit for the announcement with no
  other context.
- tab through to the same fields and prove that you see
  a little tooltip telling you what format to enter the date and time
in.

Change-Id: I006f56b8e46c3ba209e42388ca44ef054e47f73c
Reviewed-on: https://gerrit.instructure.com/56474
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>