spec: expose default sinon sandbox stubs in setup/teardown
We use qunit with sinon for javascript testing, and there's a small glue
script to tie them together that wrapped the qunit test function in
a default sinon sandbox, so that this.stub, this.spy, and this.mock were
available and automatically cleaned up.  This wasn't heavily used, at
least partly because this default sandbox wasn't available in the module
setup/teardown methods.

Through looking at github issues in the original repository
(https://github.com/cjohansen/sinon-qunit), I found a link to someone
who had run into a similar problem and solved it by instead injecting
the sinon sandbox into the test environment in the module setup
function, and then tearing it down in the teardown method (see
https://github.com/jfromaniello/jmail/blob/master/scripts/Tests/
sinon-qunit-1.0.0.js).

This almost worked right out of the box, but it needed two small
modifications. One was to modify the existing test environment by
wrapping the setup/teardown instead of injecting a new test environment,
so that other methods in the test env were maintained. The other was
that injecting into `this` wasn't working properly, causing this.sandbox
to not be available.

test plan: js specs should run and pass

Change-Id: Id91c44f29c3284790f462dc670b1c2ba53ab946f
Reviewed-on: https://gerrit.instructure.com/56567
Tested-by: Jenkins
Reviewed-by: Mark Severson <markse@instructure.com>
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>
QA-Review: Simon Williams <simon@instructure.com>