Academic Benchmark API pass error messages back to caller
Right now the API is just printing the error messages to the console and
not passing them back to the calling code.  Well, now that the calling
code is actually an API endpoint, we need that information back.

This change raises an exception with the error message now instead of
just printing to the console and returning an empty array.

Also protect against a 500 error when dereferencing an empty plugin
object

Fixes CNVS-21780

Test Plan:

    1.  Remove the API key.  Attempt to schedule an importing job
    through the API.  See that you get back an error message describing
    the lack of API key.  Restore the key for the next step.
    2.  Remove the academic_benchmark_migration_user_id settings.
    Attempt to schedule an importing job.  See that you get back an
    error message describing the lack of User ID.  Restore the user ID
    for the next step.
    3.  Set the user ID to an invalid user ID.  Attempt to schedule an
    importing job.  See that you get back an error message describing
    the inability to find the user.

Change-Id: Ibce043e8f5c15d92d8f53e142a5eee1aaa47419b
Reviewed-on: https://gerrit.instructure.com/58031
Tested-by: Jenkins
Reviewed-by: Ryan Taylor <rtaylor@instructure.com>
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Product-Review: Benjamin Porter <bporter@instructure.com>