flash an error when a grading period has an empty date
fixes CNVS-21862

This has to do with creating Grading Periods
(/courses/:id/grading_standards)

Before this commit, when a user entered an invalid date (e.g. "Octember
32") and pressed the "Save" button, a red error banner would appear at
the top of the screen, and the previously-entered valid date would be
restored to the text field when the page was reloaded.

This commit makes this same behavior occur when a user enters an empty
string into one of the date fields.

Test plan:
 * Enable the Multiple Grading Periods feature flag
 * Create a Grading Period
 * Verify that deleting the text in either date input field behaves
   identically to entering nonsense text into the date field

Change-Id: I400c5de5d413330846ecef36dca5d0ad2e145465
Reviewed-on: https://gerrit.instructure.com/58623
Reviewed-by: Cody Poll <cpoll@instructure.com>
Tested-by: Jenkins
QA-Review: Jason Carter <jcarter@instructure.com>
Product-Review: Christi Wruck