Theme Editor: Fix threading issues in preview progress updates
fixes: CNVS-22185

Because the the update_progress callback from the 
upload_to_s3 callback was happening inside a 
Parallel.each thread, we'd get random errors where 
it ran out of active record postgres connections. 
eg:
ActiveRecord::ConnectionTimeoutError
could not obtain a database connection within 5 
seconds (waited 5.000153379 seconds). The max pool 
size is currently 5; consider increasing it.

the fix is to call the callback on the main thread so 
any db queries are happening from the normal rails thread.

also, this adds a progress bar so if you run this from 
the command line, you'll get a nice progress bar like 
https://files.slack.com/files-pri/T028ZAGUD-F0896CFR9/screen_shot_2015-07-28_at_10.19.36_am.png 
so that when deployers run the:
`rake brand_configs:generate_and_upload_all` 
task, which takes a long time, they have an idea 
of how long it is going to take

test plan:

note:this is something that only broke on beta and 
not on local machines so to test that this fixes it,
you kinda need to just run it there:

* open theme editor
* make a change
* hit preview
* it should not say there was an error

Change-Id: Ife052ab9573aa677d49d8e5e6c637b0faa2b116e
Reviewed-on: https://gerrit.instructure.com/59407
Reviewed-by: Mike Nomitch <mnomitch@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Tested-by: Jenkins
Product-Review: Ryan Shaw <ryan@instructure.com>