remove k12 css variant since we don't use it for now
closes: CNVS-22285

since we took out all usage of the $is-k12 sass variable
when we removed the legacy k12 stuff we did, there
is no actual difference between the new_styles
and k12 variants. so we don't need to generate
2 different files for now (since their output will
be exactly the same).

this will cut the css generation time in half
when deploying and when generating a preview in 
theme editor since instead of outputting 2 css files
(1 for new_styles_normal_contrast and 1 for 
k12_normal_contrast) for every css bundle for every
theme, we just make one (new_styles_normal_contrast)

In the future, ui/ux might want to reintroduce
a sass variable like $is-k12 to do different things
if they are a k12 school. But when/if we do that,
we can just save it as part of the variables saved
in the school's brand_config from the theme editor.

that will be possible because when you turn on the 
k12 feature flag, it defaults to using the 
predefined 'K12 Theme' brand_config so *all* k12 
schools will have a brand_config

test plan:
make sure canvas looks the same as it did before if
you have the k12 feature flag turned on

Change-Id: I197c30b9c5ed5150a3683f836008cbdbce78b09d
Reviewed-on: https://gerrit.instructure.com/59743
Tested-by: Jenkins
Reviewed-by: Colleen Palmer <colleen@instructure.com>
Product-Review: Colleen Palmer <colleen@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>