invalid characters in initial grade do not alter workflow_state
Entering any invalid character—e.g. &, }}, {{+, or space—as an initial
grade in either the Gradebook or SpeedGrader causes the grade to be
set to null and the workflow_state to remain in unsubmitted. Because
of this, results in SpeedGrader do not show the student as needing to
be graded.

fixes CNVS-20823

test plan:
- Create a course with multiple students and an assignment of any type.
- Enter any non-grade in either Gradebook or SpeedGrader (characters
  like &, ?, +, etc…), and press 'Enter'
- Note that the character is removed as the field updates
- View the submission in SpeedGrader and note that the student shows
  as unsubmitted by the lack of orange dot nor green checkmark next to
  their name

Change-Id: Idae8c025739fdb051ca7c8089f3cc25fe1424690
Reviewed-on: https://gerrit.instructure.com/60851
Tested-by: Jenkins
Reviewed-by: Josh Simpson <jsimpson@instructure.com>
QA-Review: Jason Carter <jcarter@instructure.com>
Product-Review: Josh Simpson <jsimpson@instructure.com>