gradebook individual view doesn't show no submission as resubmitted
In the gradebook individual view's grading section, if a student has
no submission the helper message of "This assignment has been
resubmitted since it was graded last" does not appear.  Because of
this, there will not be confusion by a teacher if an assignment was
submitted or not.

fixes CNVS-21691

test plan:
- Create and publish a course with two students
- Create a student group and add both students to it
- Create and publish an assignment using the blue '+ Assignment' button
- Click on the 'Grades' link in the left column to go to the Gradebook
- Click on the 'Switch to Individual View' link in the top right
- Select one of the students
- Select the assignment
- Observe that "This assignment has been resubmitted since it was
  graded last" does not appear

Change-Id: I06f0dcb8d0e37d5150e1b20efee6d08143226017
Reviewed-on: https://gerrit.instructure.com/61839
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Amber Taniuchi <amber@instructure.com>
Tested-by: Jenkins
Product-Review: Josh Simpson <jsimpson@instructure.com>