Limit progress updates to 99%
Progress is stored as an integer but we were sending a float here, this
caused large imports to report 100% progress when they could have been
as little as 99.5% complete. For small batches this isn't an issue but
for very large customers this 1/2 percent could be many minutes to
hours.

Fixes: SIS-1403
Change-Id: I9fdcd3133e26a860c85a239652127184c179eab2
Reviewed-on: https://gerrit.instructure.com/62518
Reviewed-by: Rob Orton <rob@instructure.com>
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Tyler Pickett <tpickett@instructure.com>