Fix ui bug causing all numerical fields to show
During initialization of a numerical answer, the change event was being
fired on an element not yet inserted into the DOM. Thus the event was
not bubbling up to $(document) in order to be caught by the delegated
handler.

Moved the handler function to higher scope so as to be visible when
initializing the answer elements and modified it to accept a jquery
object.

Fixes CNVS-1604

Test plan:
 - Changing numerical answer type should show and hide the related
 fields correctly.
 - When a numerical question is saved, reloaded, and then edited, only
 the relevent numerical answer fields should be displayed for each
 numerical answer type.

Change-Id: Ia104790fe7acfd82e2e1a3934c09715feb6e3787
Reviewed-on: https://gerrit.instructure.com/64067
Reviewed-by: Matt Berns <mberns@instructure.com>
QA-Review: Michael Hargiss <mhargiss@instructure.com>
Tested-by: Jenkins
Product-Review: Jason Sparks <jsparks@instructure.com>