handle focus when adding/removing due date sets
when adding, focus goes to the input field of the new set.  when
removing, focus goes to the input field of the set that "slides in" to
the place of the removed set (so with 4 sets, deleting the original
first set, focus goes to the new first set), or if there is no new set
(when you delete the last set), focus goes to the previous set.

fixes CNVS-23043

test plan:
- on the assignment new/edit page
- add and remove due date sets with the keyboard / screenreader
- it should handle focus well, see above for the logic
- try lots of combinations, like deleting the first, last, and middle
  sets

Change-Id: Id1b7da84fddea1fe3f87d0874acf61a4978d73e0
Reviewed-on: https://gerrit.instructure.com/64837
Tested-by: Jenkins
Reviewed-by: John Corrigan <jcorrigan@instructure.com>
Reviewed-by: Mike Nomitch <mnomitch@instructure.com>
QA-Review: Landon Holmstead <lholmstead@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>