add assignment override logic to gb column sort
Gradebook does not consider assignment overrides
when sorting columns by due_date. This is updating
the sorting logic to match the method that was
written for React gradebook.

TODO: change test plan

closes CNVS-24691

Test Plan:
1. Create 4 assignments, named A, B, C, D so you can
Verify: Assignments are alphabetically sorted correctly
in gb2.
2. Give a due date to each assignment so they are due
one after another starting with assignment A (i.e. A is due_at Dec 1,
B is due_at Dec 2, etc).
Verify: Assignments are sorted correctly by due_date in gb2.
3. Change the due date for A, so that it has an override
date for Section 1 that is later than all the other assignments.
Make sure to remove the 'everyone' due date.
Verify: Column for A is now to the far right of gb2 because
it is has the latest due date. (Columns should now be
ordered B, C, D, A).
4. Change the due date for B so that it has an an override
date for both Section 1 and Section 2. Make sure to remove
the 'everyone' due date.
Verify: Assignment B is now the farthest to the right in gb2.
Assignments with multiple overrides are designed to be pushed
to the end of the line. (Columns should now be ordered C, D,
A, B).
5. Turn on gradebook performance and make sure the columns
are ordered the same.

Change-Id: I387e4b1ff69e8b6542d837f6b2d6b7b466439d78
Reviewed-on: https://gerrit.instructure.com/67126
Tested-by: Jenkins
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Jason Carter <jcarter@instructure.com>
Product-Review: Christi Wruck
Reviewed-by: Derek Bender <djbender@instructure.com>