fix speedgrader dropdown when tabbing
fixes CNVS-25145

When viewing speedgrader and doing a moderated grade, tabbing to the
students dropdown would automatically refresh the page and take you to
the first reviewer rather than stay on the same page and wait for a
student selection. This fixes that by scoping an weird accessiblity
hack for firefox to only be applied when pressing the up, down, left
and right arrow keys. Note that the tab is not included in the key
event.

Test Plan
Set up an assignment that has been submitted with moderated grading
enabled.
Have two reviewers grade the assignment
Go to speedgrader for assignment, make sure the 2nd reviewer tab is
selected.
Tab forwards (or backwards, depending on where you are on the page)
until you are on the "student 1...2..etc) dropdown (top left corner)
The page should NOT refresh when you focus on the dropdown
The page SHOULD update when a new selection is made.
This should be tested on firefox and chrome.

Change-Id: I863bcd5f5dfdbeed97d4482c594e687e291ece57
Reviewed-on: https://gerrit.instructure.com/67737
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>