spec: stabilize a selenium spec
refs CNVS-25916

the "click" function was working (in the browser)
but was returning some value that caused the
"keep trying" loop to go again, which meant
that it would keep trying to find and click
the link after the action was complete.

This makes the workflow stable and passes reliably.

Change-Id: I3a4018549dee201f5d434856897e3295f031a0d5
Reviewed-on: https://gerrit.instructure.com/69233
Reviewed-by: Rob Orton <rob@instructure.com>
Tested-by: Jenkins
Product-Review: Ethan Vizitei <evizitei@instructure.com>
QA-Review: Ethan Vizitei <evizitei@instructure.com>