load custom css/js for high contrast in newUI
In new UI, load custom CSS / JS even for users
that have High Contrast turned on.

closes: CNVS-22656

Originally, we tried to just say that to ensure schools
could not mess up their own high contrast compliance,
we would not load custom css or js for high contrast
users in NewUI.

This was problematic for things like the school using
js to add a custom help menu or something that the rest
of their classmates would see but the high contrast
user would not see. or using some "display: none"s in
css to hide things that they didn't want people to see,
but the high contrast users would still see.

This change makes it so custom css and js overrides
are loaded for high contrast users too. that means that
there is a possibility that the school could mess up
high contrast ratios with their css but I think given
the problems above this is the best we can do.

Product reviewers: (colleen, dana, aaron) please
review this description and + or - 1 if you agree.

test plan:
* turn on high contrast in New UI
* any custom css/js file you have set for your account
  (or any of your account's parents) should be loaded

Change-Id: I72ed06688f0b7b131ec00c2ea6031597277e7445
Reviewed-on: https://gerrit.instructure.com/69941
Tested-by: Jenkins
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Reviewed-by: Rob Orton <rob@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Dana Danger <dana@instructure.com>