spec: stabilize profile_spec
refs CNVS-25916

this spec is checking the DB to check it's efficacy, which is something
hard to time right, and not correct for an acceptance test anyway.

To stabilize it, we check it's efficacy by examining what the user
can see instead (same way QA would check it), and we use the normal
selenium tools for awaiting page loads.

TEST PLAN:
 1) profile_spec failure rate should drop

Change-Id: Iec1e0b95353a13d4f12b5cc005f2722d2d9f73c4
Reviewed-on: https://gerrit.instructure.com/71086
Tested-by: Jenkins
Reviewed-by: Rob Orton <rob@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>