don't swallow JS errors in production.
closes: CNVS-27680

this chunk of code effectively makes it so we
swallow all js errors in production, so you
never see the error in the web developer console.
that's kinda lame if you are trying to debug things
and figure out what went wrong.

I found this while trying to look for problems with
the webpack build. if I looked at the "console"
tab of the developer tools, nothing would show up
but if I had the js debugger turned on and open in the
"script" tab it would.

does anyone know why we do this? can you think
of a good reason to leave it there?



test plan:
* do something that causes a javascript error
  (like having the canvabadges turned on with webpack,
  sorry I can't think of an easier way to do it)
* if you have the chrome developer console open,
  you should see the error.

Change-Id: I3e8ab3d5876b48c807a22777048a25e0f7f6d522
Reviewed-on: https://gerrit.instructure.com/73240
Reviewed-by: Simon Williams <simon@instructure.com>
Tested-by: Jenkins
QA-Review: Myller de Araujo <myller@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>